+++ src/main/java/net/ccbluex/liquidbounce/features/module/modules/movement/Parkour.kt
@@ -15,13 +15,13 @@ import net.ccbluex.liquidbounce.utils.simulation.SimulatedPlayer
 object Parkour : Module("Parkour", Category.MOVEMENT, subjective = true, gameDetecting = false) {
 
     val onMovementInput = handler<MovementInputEvent> { event ->
-        val thePlayer = mc.thePlayer ?: return@handler
+        val player = mc.thePlayer ?: return@handler
 
         val simPlayer = SimulatedPlayer.fromClientPlayer(event.originalInput)
 
         simPlayer.tick()
 
-        if (thePlayer.isMoving && thePlayer.onGround && !thePlayer.isSneaking && !mc.gameSettings.keyBindSneak.isKeyDown && !simPlayer.onGround) {
+        if (player.isMoving && player.onGround && !player.isSneaking && !mc.gameSettings.keyBindSneak.isKeyDown && !simPlayer.onGround) {
             event.originalInput.jump = true
         }