+++ src/main/java/net/ccbluex/liquidbounce/features/module/modules/render/ESP.kt
@@ -41,14 +41,14 @@ object ESP : Module("ESP", Category.RENDER) {
         arrayOf("Box", "OtherBox", "WireFrame", "2D", "Real2D", "Gaussian", "Outline", "Glow"), "Box"
     )
 
-    val outlineWidth by float("Outline-Width", 3f, 0.5f..5f) { mode == "Outline" }
+    val outlineWidth by float("OutlineWidth", 3f, 0.5f..5f) { mode == "Outline" }
 
-    val wireframeWidth by float("WireFrame-Width", 2f, 0.5f..5f) { mode == "WireFrame" }
+    val wireframeWidth by float("WireFrameWidth", 2f, 0.5f..5f) { mode == "WireFrame" }
 
-    private val glowRenderScale by float("Glow-Renderscale", 1f, 0.5f..2f) { mode == "Glow" }
-    private val glowRadius by int("Glow-Radius", 4, 1..5) { mode == "Glow" }
-    private val glowFade by int("Glow-Fade", 10, 0..30) { mode == "Glow" }
-    private val glowTargetAlpha by float("Glow-Target-Alpha", 0f, 0f..1f) { mode == "Glow" }
+    private val glowRenderScale by float("GlowRenderscale", 1f, 0.5f..2f) { mode == "Glow" }
+    private val glowRadius by int("GlowRadius", 4, 1..5) { mode == "Glow" }
+    private val glowFade by int("GlowFade", 10, 0..30) { mode == "Glow" }
+    private val glowTargetAlpha by float("GlowTargetAlpha", 0f, 0f..1f) { mode == "Glow" }
 
     private val espColor = ColorSettingsInteger(this, "ESPColor").with(255, 255, 255)