+++ src/main/java/net/ccbluex/liquidbounce/features/module/modules/render/ProphuntESP.kt
@@ -25,10 +25,10 @@ import kotlin.math.pow
 
 object ProphuntESP : Module("ProphuntESP", Category.RENDER, gameDetecting = false) {
     private val mode by choices("Mode", arrayOf("Box", "OtherBox", "Glow"), "OtherBox")
-    private val glowRenderScale by float("Glow-Renderscale", 1f, 0.5f..2f) { mode == "Glow" }
-    private val glowRadius by int("Glow-Radius", 4, 1..5) { mode == "Glow" }
-    private val glowFade by int("Glow-Fade", 10, 0..30) { mode == "Glow" }
-    private val glowTargetAlpha by float("Glow-Target-Alpha", 0f, 0f..1f) { mode == "Glow" }
+    private val glowRenderScale by float("GlowRenderscale", 1f, 0.5f..2f) { mode == "Glow" }
+    private val glowRadius by int("GlowRadius", 4, 1..5) { mode == "Glow" }
+    private val glowFade by int("GlowFade", 10, 0..30) { mode == "Glow" }
+    private val glowTargetAlpha by float("GlowTargetAlpha", 0f, 0f..1f) { mode == "Glow" }
 
     private val color by color("Color", Color(0, 90, 255))