+++ src/main/java/net/ccbluex/liquidbounce/features/module/modules/render/StorageESP.kt
@@ -35,12 +35,12 @@ object StorageESP : Module("StorageESP", Category.RENDER) {
     private val mode by
     choices("Mode", arrayOf("Box", "OtherBox", "Outline", "Glow", "2D", "WireFrame"), "Outline")
 
-    private val glowRenderScale by float("Glow-Renderscale", 1f, 0.5f..2f) { mode == "Glow" }
-    private val glowRadius by int("Glow-Radius", 4, 1..5) { mode == "Glow" }
-    private val glowFade by int("Glow-Fade", 10, 0..30) { mode == "Glow" }
-    private val glowTargetAlpha by float("Glow-Target-Alpha", 0f, 0f..1f) { mode == "Glow" }
+    private val glowRenderScale by float("GlowRenderscale", 1f, 0.5f..2f) { mode == "Glow" }
+    private val glowRadius by int("GlowRadius", 4, 1..5) { mode == "Glow" }
+    private val glowFade by int("GlowFade", 10, 0..30) { mode == "Glow" }
+    private val glowTargetAlpha by float("GlowTargetAlpha", 0f, 0f..1f) { mode == "Glow" }
 
-    private val espColorMode by choices("ESP-ColorMode", arrayOf("None", "Custom"), "None")
+    private val espColorMode by choices("ESPColorMode", arrayOf("None", "Custom"), "None")
     private val espColor = ColorSettingsInteger(this, "ESPColor")
     { espColorMode == "Custom" }.with(255, 179, 72)